Add missing parent tasks, linting, remove mockdata code
Added the ability to make API calls to JIRA to get more data. Currently this is used for fetching missing parent info when the query returns a subtask, but not the parent. Added missing eslint deps, applied formatting. Removed code for mock data. Use a src in the script tag to point to the hot-reload .js file while developing.